+++ packages/react/src/slider/indicator/SliderIndicator.tsx
@@ -1,11 +1,32 @@
 'use client';
 import * as React from 'react';
-import type { BaseUIComponentProps } from '../../utils/types';
-import { useComponentRenderer } from '../../utils/useComponentRenderer';
+import type { BaseUIComponentProps, Orientation } from '../../utils/types';
+import { useRenderElement } from '../../utils/useRenderElement';
 import { useSliderRootContext } from '../root/SliderRootContext';
 import { sliderStyleHookMapping } from '../root/styleHooks';
-import type { SliderRoot } from '../root/SliderRoot';
-import { useSliderIndicator } from './useSliderIndicator';
+import { valueArrayToPercentages, type SliderRoot } from '../root/SliderRoot';
+
+function getRangeStyles(
+  orientation: Orientation,
+  offset: number,
+  leap: number,
+): React.CSSProperties {
+  if (orientation === 'vertical') {
+    return {
+      position: 'absolute',
+      bottom: `${offset}%`,
+      height: `${leap}%`,
+      width: 'inherit',
+    };
+  }
+
+  return {
+    position: 'relative',
+    insetInlineStart: `${offset}%`,
+    width: `${leap}%`,
+    height: 'inherit',
+  };
+}
 
 /**
  * Visualizes the current value of the slider.
@@ -14,28 +35,42 @@ import { useSliderIndicator } from './useSliderIndicator';
  * Documentation: [Base UI Slider](https://base-ui.com/react/components/slider)
  */
 const SliderIndicator = React.forwardRef(function SliderIndicator(
-  props: SliderIndicator.Props,
-  forwardedRef: React.ForwardedRef<HTMLElement>,
+  componentProps: SliderIndicator.Props,
+  forwardedRef: React.ForwardedRef<HTMLDivElement>,
 ) {
-  const { render, className, ...otherProps } = props;
+  const { render, className, ...elementProps } = componentProps;
 
-  const { disabled, max, min, orientation, state, values } = useSliderRootContext();
+  const { max, min, orientation, state, values } = useSliderRootContext();
 
-  const { getRootProps } = useSliderIndicator({
-    disabled,
-    max,
-    min,
-    orientation,
-    values,
-  });
+  const percentageValues = valueArrayToPercentages(values.slice(), min, max);
 
-  const { renderElement } = useComponentRenderer({
-    propGetter: getRootProps,
-    render: render ?? 'div',
+  let style: React.CSSProperties;
+
+  if (percentageValues.length > 1) {
+    const trackOffset = percentageValues[0];
+    const trackLeap = percentageValues[percentageValues.length - 1] - trackOffset;
+
+    style = getRangeStyles(orientation, trackOffset, trackLeap);
+  } else if (orientation === 'vertical') {
+    style = {
+      position: 'absolute',
+      bottom: 0,
+      height: `${percentageValues[0]}%`,
+      width: 'inherit',
+    };
+  } else {
+    style = {
+      position: 'relative',
+      insetInlineStart: 0,
+      width: `${percentageValues[0]}%`,
+      height: 'inherit',
+    };
+  }
+
+  const renderElement = useRenderElement('div', componentProps, {
     state,
-    className,
     ref: forwardedRef,
-    extraProps: otherProps,
+    props: [{ style }, elementProps],
     customStyleHookMapping: sliderStyleHookMapping,
   });
